Move From<IpAddr> impl for NetAddr from crate root into netaddr mod
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
